Fixed deadlock in game clock
The game clock was using `synchronized` to keep a basic lock on the
scheduled events list internally when processing each event during a
tick. Further, when we were scheduling events on the clock we also
were implicitly waiting on a lock via the synchronized list's add
method.

This was causing a deadlock to occur when `Runnable` for a scheduled
event attempted to itself schedule an event or an interval. Because the
processing for each event was in a loop that was synchronized on the
event list, the inner call to the scheduling methods would block the
thread indefinitely as they could never acquire a lock on the events
list in order to add a new entry.

Fixed this by using a semaphore lock with a single license. Now when
the main processing loop for a tick runs it waits until it can acquire
a scheduling lock, after which it proceeds with processing each event
after which the lock is released.

Then I changed the scheduling methods to "try" to get a lock, if they
couldn't then they now use a backup queue for all events that were
scheduled during the processing of clock events.

Finally after main event processing loop another loop runs over the
schedule queue to ensure all events that were pending are added on to
the master clock schedule list.

I remember running into this issue a couple years ago, but I managed to
work around it by changing the implementation of whatever the hell I
was working on at the time. My guess is that it was for buff timers on
the `AbstractPlayer` class, but even so I do like that impelementation
because it just uses dates to determine the number of seconds left as
opposed to updating some threaded counter.