Changed alteration design
Removed sequencer and added truesight. Sequencer on second thought was way too powerful of a spell when used in combination with evocation spells like flamestrike (the 75% discount alone, not to mention the massive amounts of damage when stacking huge spells like that).

Further we really needed a way for characters to have the option of detecting invisible game entities. While this traditionally belongs in a divination school, there really isn't a place for a skill like that as of yet, so I decided alteration was the best bet (as it already had a focus on buffing).