Tweaked Evocation Earth Bombs
The design as written didn't make much sense as quake was comboed off of landslide but the player wouldn't have enough MP to actually perform the combo. Further, the bombs felt too vanilla and similar to the fire skills (with the exception that they added debuffs).

Made two changes based on feedback from @alieander:
1. Quake now combos when landslide is on cooldown.
2. Both quake and landslide use all remaining MP and scale damage based on what percentage of the maximum was used.